introduce the couch_hooks module
Hooks are function attached to nodes events. There are for now 2 kind of node
events: global or related to a database.

You can hook a function to a global event using couch_hooks/4 and to a
database using couch_hooks/5. Not that passing the atom all instead of a
database name make the function react to all databases events.

For now 2 databases events have been created:

db_updated(DbName, Event)
ddoc_updated(DbName, DDocId)

While we are here publish database events to a pid inside a couch_server hook
instead of using a gen_server.