extract couch_js code and other module from couch
couch_js and couch_os_processes are now available as external OTP libs so we
can make it optionnal if needed. While I was here I also created a new lib
`couch_extra` where all extra feature of couchdb not relevant to couchapps or
db/indexes will be put.

barrel-couchjs is a dependency of couch for now